Refactor: Moved all extension methods to Extensions namespace
