+++ src/QueryFramework.Core.Tests/Builders/QueryConditionBuilderTests.cs
@@ -2,7 +2,7 @@
 using FluentAssertions;
 using Moq;
 using QueryFramework.Abstractions;
-using QueryFramework.Abstractions.Extensions.Builders;
+using QueryFramework.Abstractions.Extensions;
 using QueryFramework.Core.Builders;
 using Xunit;
