+++ src/QueryFramework.QueryParsers/SingleEntityQueryParser.cs
@@ -4,7 +4,7 @@ using System.Globalization;
 using System.Linq;
 using QueryFramework.Abstractions;
 using QueryFramework.Abstractions.Builders;
-using QueryFramework.Abstractions.Extensions.Builders;
+using QueryFramework.Abstractions.Extensions;
 using QueryFramework.Abstractions.Queries.Builders;
 using QueryFramework.Core.Builders;
 using QueryFramework.QueryParsers.Extensions;
