+++ src/QueryFramework.SqlServer.Tests/IntegrationTests.cs
@@ -6,7 +6,6 @@ using FluentAssertions;
 using Moq;
 using QueryFramework.Core.Extensions;
 using QueryFramework.Core.Queries.Builders;
-using QueryFramework.Core.Queries.Builders.Extensions;
 using QueryFramework.SqlServer.Tests.Repositories;
 using Xunit;
