+++ src/QueryFramework.SqlServer.Tests/QueryProcessorTests.cs
@@ -11,7 +11,6 @@ using QueryFramework.Abstractions.Queries;
 using QueryFramework.Core.Extensions;
 using QueryFramework.Core.Queries;
 using QueryFramework.Core.Queries.Builders;
-using QueryFramework.Core.Queries.Builders.Extensions;
 using QueryFramework.SqlServer.Tests.TestHelpers;
 using Xunit;
